pyproject.toml
[project]
name = "agentflow"
version = "1.0.1"
description = "Real-time data platform serving context to AI agents"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "bcrypt>=5,<6",
    "confluent-kafka>=2.5,<3",
    "fastapi>=0.111,<1",
    "uvicorn[standard]>=0.30,<1",
    "dagster>=1.7,<2",
    "duckdb>=1.1,<2",
    "httpx>=0.27,<1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.41,<2",
    "opentelemetry-instrumentation-fastapi>=0.62b0,<1",
    "opentelemetry-instrumentation-httpx>=0.62b0,<1",
    "opentelemetry-sdk>=1.41,<2",
    "pandera>=0.20,<1",
    "prometheus-client>=0.21,<1",
    "pyarrow>=17,<19",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.5,<3",
    "pyyaml>=6,<7",
    "sqlglot>=30,<31",
    "structlog>=24.4,<26",
]

[project.optional-dependencies]
flink = [
    "apache-flink==1.19.1",
]
cloud = [
    "boto3>=1.35,<2",
    "pyiceberg>=0.7,<1",
]
llm = [
    "anthropic>=0.39,<1",
]
load = [
    "locust>=2.29,<3",
]
integrations = [
    "langchain>=0.3,<1",
    "llama-index-core>=0.12,<1",
]
dev = [
    "bandit>=1.9,<2",
    "hypothesis>=6,<7",
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5,<6",
    "testcontainers[kafka]>=4.9,<5",
    "ruff>=0.6,<1",
    "mypy>=1.11,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PT"]
ignore = ["S101", "S311", "S608"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S603",
    "S607",
    "S310",
    "S104",
    "N802",
    "E402",
    "B017",
    "PT011",
    "A002",
    "E501",
]
"src/serving/backends/clickhouse_backend.py" = [
    "S310",
    "E501",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring Docker services",
    "requires_docker: marks tests that require local Docker",
]
asyncio_mode = "auto"

[tool.mutmut]
paths_to_mutate = [
    "src/serving/api/auth.py",
    "src/serving/masking.py",
    "src/serving/semantic_layer/query_engine.py",
    "src/processing/outbox.py",
    "src/serving/api/rate_limiter.py",
]
also_copy = [
    "src",
    "config",
    "sdk",
]
tests_dir = [
    "tests/unit",
    "tests/property",
    "tests/integration/test_outbox.py",
]
mutate_only_covered_lines = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "src.processing.flink_jobs.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "src.serving.semantic_layer.query.*"
disable_error_code = ["attr-defined"]
